change ABI back to 64 bit

+++ sdk/core/src/internal/registers.rs
@@ -1,12 +1,12 @@
 //! Definitions of special registers
 //!
 //! We use registers to share data between host and guest.
-//! Each register is defined by a unique register-id (which is a `u32`) and can hold an arbitrary amount of bytes.
+//! Each register is defined by a unique register-id (which is a `u64`) and can hold an arbitrary amount of bytes.
 //!
 //! It is generally assumed, that the guest "knows" how to handle the bytes of the register.
 
 /// Register used for atomic operations
-pub(crate) const REGISTER_ATOMIC_OP: u32 = u32::MAX - 1;
+pub(crate) const REGISTER_ATOMIC_OP: u64 = u64::MAX - 1;
 
 /// Register used to feed input into the contract
-pub const REGISTER_INPUT: u32 = 0;
+pub const REGISTER_INPUT: u64 = 0;
